Started a refactor
